requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
coverage==7.6.7
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.1.0
filters==1.3.2
flake8==7.1.1
idna==3.10
inflection==0.5.1
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13